Move views to gem's app/views/ directory
This enables people to try Quo Vadis without having to copy the views into
their app.  If they go on to adopt the gem, they will still be expected
to copy the views into their app (as per the installation instructions)
to customise them to their look and feel.

See #22.